feature(select-component): rename content view
